exercism

+++ etc/init/linux/exercism.sh
@@ -1,7 +1,5 @@
 #!/bin/bash
 
-echo "install exercism"
-
 # Stop script if errors occur
 trap 'echo Error: $0:$LINENO stopped; exit 1' ERR INT
 set -eu
@@ -11,7 +9,12 @@ set -eu
 # For more information, see etc/README.md
 . "$DOTPATH"/etc/lib/vital.sh
 
-wget https://github.com/exercism/cli/releases/download/v2.3.0/exercism-linux-64bit.tgz
-tar -xzvf exercism-linux-64bit.tgz 
-mv ./exercism ~/bin/
-rm ./exercism-linux-64bit.tgz 
+if is_exists "exercism"; then
+  echo "exercism is installed"
+else
+  echo "install exercism"
+  wget https://github.com/exercism/cli/releases/download/v2.3.0/exercism-linux-64bit.tgz
+  tar -xzvf exercism-linux-64bit.tgz 
+  mv ./exercism ~/bin/
+  rm ./exercism-linux-64bit.tgz 
+fi
